Safari-609.1.11.1/LayoutTests/canvas/philip/tests/2d.composite.globalComposite-expected.txt

Test canvas globalCompositeOperator
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
PASS ctx.globalCompositeOperation is "clear"
PASS ctx.globalCompositeOperation is "copy"
PASS ctx.globalCompositeOperation is "source-over"
PASS ctx.globalCompositeOperation is "destination-over"
PASS ctx.globalCompositeOperation is "source-in"
PASS ctx.globalCompositeOperation is "destination-in"
PASS ctx.globalCompositeOperation is "source-out"
PASS ctx.globalCompositeOperation is "destination-out"
PASS ctx.globalCompositeOperation is "source-atop"
PASS ctx.globalCompositeOperation is "destination-atop"
PASS ctx.globalCompositeOperation is "xor"
PASS ctx.globalCompositeOperation is "lighter"
PASS ctx.globalCompositeOperation is "multiply"
PASS ctx.globalCompositeOperation is "screen"
PASS ctx.globalCompositeOperation is "overlay"
PASS ctx.globalCompositeOperation is "darken"
PASS ctx.globalCompositeOperation is "lighten"
PASS ctx.globalCompositeOperation is "color-dodge"
PASS ctx.globalCompositeOperation is "color-burn"
PASS ctx.globalCompositeOperation is "hard-light"
PASS ctx.globalCompositeOperation is "soft-light"
PASS ctx.globalCompositeOperation is "difference"
PASS ctx.globalCompositeOperation is "exclusion"
PASS ctx.globalCompositeOperation is "hue"
PASS ctx.globalCompositeOperation is "saturation"
PASS ctx.globalCompositeOperation is "color"
PASS ctx.globalCompositeOperation is "luminosity"
PASS successfullyParsed is true

TEST COMPLETE
